backend/requirements.txt
django>=5.0
djangorestframework>=3.15
django-environ>=0.11
django-cors-headers>=4.3
psycopg2-binary>=2.9
celery>=5.3
redis>=5.0
httpx>=0.27
beautifulsoup4>=4.12
lxml>=5.2
readability-lxml>=0.8
playwright>=1.46
tenacity>=9.0
pydantic>=2.8
cryptography>=42.0
python-slugify>=8.0
python-dateutil>=2.9
uvicorn>=0.30
